+++ aldryn_newsblog/views.py
@@ -467,8 +467,9 @@ class RelatedArticles(View):
         if not self.has_view_or_change_permission(request, article):
             return JsonResponse(data, status=403)
 
-        qs = Article.objects.values_list('pk', 'translations__title').filter(app_config=app_config)
+        qs = Article.objects.values_list(
+            'pk', 'translations__title', 'publishing_date').filter(app_config=app_config).order_by('pk').distinct('pk')
         if article_id is not None:
             qs = qs.exclude(pk__in=(article.pk, ) + tuple(article.related.values_list('pk', flat=True)))
-        data["articles"] = tuple(qs)
+        data["articles"] = [(opt[0], opt[1]) for opt in sorted(qs, key=lambda item: item[2], reverse=True)]
         return JsonResponse(data)